Enhance payment, scheduling, and management functionalities (#15)
* feat: update services with new payment and schedule functionalities

* Nam (#8)

* feat: implement exam management system

* feat: rename routes for student portals

---------


* feat: update services with new payment and schedule functionalities


* check

* Thêm báo cáo và lịch sử chi tiết

* feat: refactor codebase

* feat: enhance cashbook management with role mapping and data structure updates

* Refactor HomePage to include FinanceSummary component and remove unused transaction and debtor data

* feat: add TeacherScheduleToday component and integrate with FinancialReport

* feat: update EmailService to use createTransport for SMTP configuration

* feat: enhance UI for student/teacher portal

* feat: implement yearly cashbook summary endpoint and update related services and components

* feat: change create class function

* feat: enhance permission management and error handling across various modules

* update cashbook service

* feat: Implement Genetic Algorithm for schedule generation

- Add GeneticAlgorithmService to handle the genetic algorithm logic for scheduling classes.
- Introduce helper functions for data preparation, fitness calculation, crossover, and mutation.
- Integrate the genetic algorithm into the ScheduleService for auto-generating schedules based on class requests.
- Create ScheduleRequestService to manage schedule requests, including initialization, updating preferences, and fetching requests.
- Define new types for class requests, genetic algorithm inputs, and schedule requests to improve type safety and clarity.
- Add validation schemas for initializing and updating schedule requests using Zod.

* feat: improve UI class schedule creation features

* feat: implement score sheet for student portal

* feat: add schedule field to ClassRequest model and implement genetic algorithm for class scheduling

* feat: implement enrollment operations

* fix: type error

* feat: convert text to Vietnamese

* feat: enhance LoadingOverlay with step and total props, improve animations and styles

* fix cancel invoice

* Refactor AutoSchedule components for improved UI and functionality

* feat: enhance TeacherDetail and TeacherManager components with improved UI and permission checks

- Refactor TeacherDetail component to include new InfoRow and StatCard sub-components for better layout and readability.
- Add functionality to fetch schedules for the current week in TeacherDetail.
- Update loading and error states in TeacherDetail for better user experience.
- Integrate RequirePermission component in TeacherManager to restrict access to certain actions based on user permissions.
- Clean up permission constants by commenting out unused CRM permissions.
- Update user routes to require multiple permissions for accessing teacher data.
- Modify schedule service to support filtering schedules by date range.
- Adjust seed data for students to use a new email format.

* feat: add permission checks for various management pages

* feat: adjust date handling in getTodaySchedules for Vietnam timezone

* feat: optimize genetic algorithm parameters based on class size and improve logging

* feat: add MY_VIEW permission for invoices and update routes accordingly

* feat: API validations and UI components

* feat(seed): add script to create unscheduled classes for performance testing

---------

+++ server/src/services/exam.service.ts
@@ -124,23 +124,25 @@ export class ExamService {
         const exam = await ExamModel.create(data);
         const classes = await ClassModel.findById(data.classId);
         if (!classes) throw new Error('Không tìm thấy lớp học');
-        const notifications = classes.studentIds.map(studentId => {
-            return ({
-                userId: studentId,
-                title: `Lớp ${classes.name} có bài kiểm tra mới`,
-                content: `${exam.title} ${exam.description ? `- ${exam.description}` : ''}`,
-                examId: exam._id,
-                isRead: false
+        if (data.status === ExamStatus.PUBLISHED) {
+            const notifications = classes.studentIds.map(studentId => {
+                return ({
+                    userId: studentId,
+                    title: `Lớp ${classes.name} có bài kiểm tra mới`,
+                    content: `${exam.title} ${exam.description ? `- ${exam.description}` : ''}`,
+                    examId: exam._id,
+                    isRead: false
+                })
             })
-        })
-        const savedNotifs = await AttendanceNotificationModel.insertMany(notifications);
-        const io = getIO();
-        savedNotifs.forEach(notif => {
-            const socketId = userSocketMap.get(notif.userId.toString());
-            if (socketId) {
-                io.to(socketId).emit('new_notification', notif.toObject());
-            }
-        });
+            const savedNotifs = await AttendanceNotificationModel.insertMany(notifications);
+            const io = getIO();
+            savedNotifs.forEach(notif => {
+                const socketId = userSocketMap.get(notif.userId.toString());
+                if (socketId) {
+                    io.to(socketId).emit('new_notification', notif.toObject());
+                }
+            });
+        }
         return exam;
     }
     async updateExam(id: string, data: UpdateExamType) {
@@ -178,10 +180,12 @@ export class ExamService {
 
     async deleteExam(id: string) {
         try {
-            await ExamSubmissionModel.deleteMany({ examId: id });
+            const existSubmissions = await ExamSubmissionModel.find({ examId: id });
+            if (existSubmissions.some(sub => sub.status === 'IN_PROGRESS')) throw new Error('Không thể xóa bài kiểm tra vì đang có học viên làm bài');
+            await AttendanceNotificationModel.deleteMany({ examId: id });
             return ExamModel.findByIdAndDelete(id).lean();
-        } catch (error) {
-            throw new Error('Không thể xóa bài kiểm tra');
+        } catch (error: any) {
+            throw new Error(error.message || 'Không thể xóa bài kiểm tra');
         }
     }
